adjusted folder structure and adjusted nav routing
