+++ package.json
@@ -13,6 +13,7 @@
     "embla-carousel-react": "^8.3.1",
     "force": "^0.0.3",
     "framer-motion": "^10.16.12",
+    "hamburger-react": "^2.5.1",
     "mdb-react-ui-kit": "^9.0.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.10.6",
