+++ src/Components/About.css
@@ -33,6 +33,7 @@ h2 {
   justify-content: center;
   align-self: center;
   font-size: 5rem;
+  margin-left: 2rem;
 }
 
 @keyframes wave-animation {
@@ -102,6 +103,7 @@ h2 {
   }
   .aboutHeader {
     align-self: center;
-    font-size: 4rem;
+    font-size: 3rem;
+    margin: 0;
   }
 }
