--- src/Hamburger.js
@@ -1,39 +1,0 @@
-const Hamburger = () => (
-  <svg
-    xmlns='http://www.w3.org/2000/svg'
-    width='52'
-    height='24'
-    viewBox='0 0 52 24'
-  >
-    <g id='Group_9' data-name='Group 9' transform='translate(-294 -47)'>
-      <rect
-        id='Rectangle_3'
-        data-name='Rectangle 3'
-        width='42'
-        height='4'
-        rx='2'
-        transform='translate(304 47)'
-        fill='#574c4c'
-      />
-      <rect
-        id='Rectangle_5'
-        data-name='Rectangle 5'
-        width='42'
-        height='4'
-        rx='2'
-        transform='translate(304 67)'
-        fill='#574c4c'
-      />
-      <rect
-        id='Rectangle_4'
-        data-name='Rectangle 4'
-        width='52'
-        height='4'
-        rx='2'
-        transform='translate(294 57)'
-        fill='#574c4c'
-      />
-    </g>
-  </svg>
-);
-export default Hamburger;
