+++ src/Styles/ProjectCard.css
@@ -5,6 +5,7 @@
   margin: 2rem;
   transition: all 0.5s ease 0s !important;
   box-shadow: 0px 0px 5px 0px #0ff !important;
+  margin: 2rem;
 }
 .customCard:hover {
   transform: scale(1.02);
