+++ src/Styles/Projects.css
@@ -12,10 +12,11 @@
   margin-top: 5rem;
   background-color: rgb(101, 109, 112, 0.3);
   height: 60rem;
-  width: 80%;
-  box-shadow: inset 0px 0px 6px rgba(0, 0, 0, 0.9);
-  border-radius: 1rem;
-
+  max-width: fit-content;
+  box-shadow: inset 0px 0px 7px rgba(0, 0, 0, 0.9);
+  border-radius: 2rem;
+  padding-left: 2rem;
+  padding-right: 2rem;
   justify-content: center;
 }
 
@@ -55,8 +56,10 @@ p {
     flex-direction: column;
     align-items: center;
     justify-content: center;
+    padding: 0;
   }
   .customCard {
+    align-self: center;
     align-self: center;
     width: 90%;
     margin-left: 1.2rem;
