+++ src/Styles/techCarousel.css
